Add ReplyTo parameter support to Send Email Misc: Fix Email feature folder name (and namespaces) in UnitTests project.
