+++ examples/Mailtrap.Example.Email.Send/Program.cs
@@ -180,6 +180,8 @@ internal sealed class Program
         request.Cc.Add(new("ursa@ursamajor.gov"));
         request.Bcc.Add(new("aliens@milkyway.net"));
 
+        request.ReplyTo = new("no-reply@earth.com");
+
         // HTML body.
         // At least one of the Text or Html body must be specified.
         request.HtmlBody =
@@ -234,9 +236,8 @@ internal sealed class Program
         // Sender (Display name is optional)
         request.From("john.doe@demomailtrap.com", "John Doe");
 
-        // Alternatively, you can set the property directly
-        request.From = new("john.doe@demomailtrap.com", "John Doe");
-
+        // Reply To
+        request.ReplyTo("info@domain.com");
 
         // You can use simple email as recipient
         request.To("hero.bill@galaxy.net");
@@ -249,7 +250,9 @@ internal sealed class Program
         request.Cc(vipEmail);
 
         // Alternatively, you could pass a collection at once
-        request.Bcc(new EmailAddress("first@domain.com"), new EmailAddress("second@domain.com"));
+        request.Bcc(
+            new EmailAddress("first@domain.com"),
+            new EmailAddress("second@domain.com"));
 
         // Subject
         request.Subject("Invitation to Earth");
