+++ src/Mailtrap.Abstractions/Emails/Validators/SendEmailRequestValidator.cs
@@ -16,6 +16,10 @@ internal sealed class SendEmailRequestValidator : AbstractValidator<SendEmailReq
     {
         RuleFor(r => r.From!).NotNull().SetValidator(EmailAddressValidator.Instance);
 
+        RuleFor(r => r.ReplyTo!)
+            .SetValidator(EmailAddressValidator.Instance)
+            .When(r => r.ReplyTo is not null);
+
         RuleFor(r => r.To).Must(r => r.Count is <= 1000);
         RuleForEach(r => r.To).SetValidator(EmailAddressValidator.Instance);
