+++ tests/Mailtrap.UnitTests/Emails/Validators/EmailAddressValidatorTests.cs
@@ -5,7 +5,7 @@
 // -----------------------------------------------------------------------
 
 
-namespace Mailtrap.UnitTests.Email.Validators;
+namespace Mailtrap.UnitTests.Emails.Validators;
 
 
 [TestFixture]
